LayoutTests/svg/animations/animate-text-nested-transforms-expected.txt

SVG 1.1 dynamic animation tests

This is some text.
test to determine whether auto-rotate animateMotion path animations pre-muliply or post-multiply animation transform matrix

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rootSVGElement.getBBox().x is almost 196 (within 20)
PASS rootSVGElement.getBBox().y is almost -186 (within 20)
PASS rootSVGElement.getBBox().x is almost 370 (within 20)
PASS rootSVGElement.getBBox().y is almost 547 (within 20)
PASS successfullyParsed is true

TEST COMPLETE
